feat: use general shebang and install missing packages

+++ node/install.sh
@@ -1,22 +1,25 @@
-#!/bin/bash
-set -e
+#!/usr/bin/env bash
+set -Eueo pipefail
 
 if [ ! -f /etc/apt/sources.list.d/node*.list ]
 then
     curl -sL https://deb.nodesource.com/setup_14.x | sudo -E bash -
 fi
 
-if [ ! $(which node) ]
-then
-    sudo apt install -y build-essential gcc g++ make nodejs
-fi
+PACKAGES=(
+  build-essential
+  gcc
+  g++
+  make
+  nodejs
+)
 
-if [ ! $(which yarn) ]
-then
-    sudo npm install -g yarn
-fi
+for pkg in "${PACKAGES[@]}"; do
+  [ ! $(dpkg -s $pkg >/dev/null 2>&1;) ] || sudo apt -y install $pkg
+done
 
-sudo npm install -g npm
-sudo npm update -g
+[ ! $(dpkg -s yarn >/dev/null 2>&1;) ] || sudo npm install -g yarn
+[ ! $(npm outdated -g npm >/dev/null 2>&1;) ] || sudo npm install -g npm
+[ ! $(npm outdated -g >/dev/null 2>&1;) ] || sudo npm update -g
 
 exit 0
